New Input System and Interaction overhaul
I've imported the new input system and replaced all of the player's actions with the new input actions in code. I added a PlayerInputHandler.cs script that takes care of all the player's inputs.

The interaction system is optimized so it doesn't call the "HideUI" function many repeated times.

Added support for keyboard and controller, with the UI sprites responding whichever is currently active. This is through the Control Scheme HUD scriptable object.

Added Controller support for navigating the menues, to do this I changed the Event System to use the new Unity Input System version.

For the Incubation pod and the shelter, now the player's movement is stopped upon interaction ONLY if we're using the controller so we can navigate the menu. If we're using keyboard then we don't stop movement since they can use the mouse to interact instead. If using keyboard and the player leaves the vicinity, then we hide the current UI they're on.