Re-organized Custom Inspector and Titles
Now all the Fields are organized in the inspector for visual elegance. I've added a sprite so we can see what the alien looks like, as well as a custom [Title()] attribute and coloured buttons!

I also now fixed the visual List so we can see all the aliens that are inside of the container visually (since we can't serialize a dictionary I create a List<string> instead)