Alien Tool: Bug Fixing and Helper Box
I fixed a null reference error where upon starting the project again the editor would make the container variable null for some reason, so now OnEnable it sets the container to it's parent asset.

I also changed the colour of the "delete" button and the "rename" button. For the rename button I added extra feedback that the designer needs to click the button as it turns yellow if the name of the asset and their information do not match.

I added a helper box warning the designer if the Tiers of the 2 aliens are not matching.