Added Pause Menu, Audio Manager, and Gene Splicing UI
Now the player can pause by pressing "Esc" or "Start" for gamepads. None of the buttons other than "Resume" and "Quit" are functional currently. An audio manager was added so we can add in music and SFX starting soon.
I also updated the Gene Splicing UI to make it bigger and removed the "+" on the buttons when an alien is added.